requirements.txt
playwright
fastapi
uvicorn[standard]
google-genai
python-dotenv
beautifulsoup4
apscheduler
jinja2
